+++ lib/jei/builders/included_node_builder.rb
@@ -4,12 +4,14 @@ module Jei
       include Nodes
 
       # @param [Set<Serializer>] serializers
+      # @param [Hash<String, String>] fieldset
       # @return [IncludedNode]
-      def self.build(serializers)
+      def self.build(serializers, fieldsets = {})
         node = IncludedNode.new
 
         serializers.each do |serializer|
-          node.children << ResourceNodeBuilder.build(serializer)
+          fieldset = fieldsets[serializer.type]
+          node.children << ResourceNodeBuilder.build(serializer, fieldset)
         end
 
         node